backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=21.0
whitenoise>=6.6
python-dotenv>=1.0
psycopg2-binary>=2.9
fpdf>=1.7.2
